+++ bootstrap.php
@@ -7,32 +7,20 @@
  * files to initiate a new Kirby site
  */
 
-// build a default roots array if not available
-if(!isset($roots)) {
-  $dir   = dirname(__FILE__);
-  $root  = dirname($dir);
-  $roots = array(
-    'root'         => $root,
-    'root.kirby'   => $dir,
-    'root.site'    => $root . DIRECTORY_SEPARATOR . 'site',
-    'root.content' => $root . DIRECTORY_SEPARATOR . 'content'
-  );
-}
-
 // global key for direct inclusion protection
-define('KIRBY', true);
+if(!defined('KIRBY')) define('KIRBY', true);
 
 // shortcut for the directory separator
-define('DS', DIRECTORY_SEPARATOR);
+if(!defined('DS')) define('DS', DIRECTORY_SEPARATOR);
 
-// build constants for all roots
-define('ROOT',                $roots['root']);
-
-// the content folder
-define('ROOT_CONTENT',        $roots['root.content']);
+// last exit to define basic roots
+if(!defined('ROOT_KIRBY'))          define('ROOT_KIRBY',         dirname(__FILE__));
+if(!defined('ROOT'))                define('ROOT',               dirname(ROOT_KIRBY));
+if(!defined('ROOT_CONTENT'))        define('ROOT_CONTENT',       ROOT . DS . 'content');
+if(!defined('ROOT_SITE'))           define('ROOT_SITE',          ROOT . DS . 'site');
+if(!defined('ROOT_KIRBY_TOOLKIT'))  define('ROOT_KIRBY_TOOLKIT', ROOT_KIRBY . DS . 'toolkit');
 
 // all stuff in the main kirby folder
-define('ROOT_KIRBY',          $roots['root.kirby']);
 define('ROOT_KIRBY_LIB',      ROOT_KIRBY . DS . 'lib');
 define('ROOT_KIRBY_LEGACY',   ROOT_KIRBY . DS . 'legacy');
 define('ROOT_KIRBY_TAGS',     ROOT_KIRBY . DS . 'tags');
@@ -41,7 +29,6 @@ define('ROOT_KIRBY_PLUGINS',  ROOT_KIRBY . DS . 'plugins');
 define('ROOT_KIRBY_MODALS',   ROOT_KIRBY . DS . 'modals');
 
 // all stuff in the site folder
-define('ROOT_SITE',           $roots['root.site']);
 define('ROOT_SITE_CACHE',     ROOT_SITE . DS . 'cache');
 define('ROOT_SITE_TEMPLATES', ROOT_SITE . DS . 'templates');
 define('ROOT_SITE_SNIPPETS',  ROOT_SITE . DS . 'snippets');
@@ -50,13 +37,20 @@ define('ROOT_SITE_PLUGINS',   ROOT_SITE . DS . 'plugins');
 define('ROOT_SITE_LANGUAGES', ROOT_SITE . DS . 'languages');
 define('ROOT_SITE_TAGS',      ROOT_SITE . DS . 'tags');
 
-// load the kirby toolkit
-include(ROOT_KIRBY_LIB . DS . 'kirby.php');
+// load the toolkit
+require_once(ROOT_KIRBY_TOOLKIT . DS . 'bootstrap.php');
 
 // autoloader for all classes in the lib
 function libLoader($class) {
+  
+  // library stuff
   $file = ROOT_KIRBY_LIB . DS . strtolower(str_replace('Kirby', '', $class)) . '.php';
-  if(file_exists($file)) require_once($file);
+  
+  if(file_exists($file)) {
+    require_once($file);
+    return;
+  } 
+
 }
 
 spl_autoload_register('libLoader');