+++ legacy/uriquery.php
@@ -1,6 +1,0 @@
-<?php
-
-// direct access protection
-if(!defined('KIRBY')) die('Direct access is not allowed');
-
-class UriQuery extends KirbyUriQuery {}