+++ test/VisitorTest.php
@@ -8,7 +8,7 @@ class VisitorTest extends PHPUnit_Framework_TestCase {
   }
   
   public function testMethods() {
-    $this->assertEquals('127.0.0.1', $this->visitor->ip());
+    $this->assertEquals('0.0.0.0', $this->visitor->ip());
     $this->assertEquals('en', $this->visitor->language()->code());
   }
 }